LayoutTests/imported/w3c/web-platform-tests/css/css-break/inheritance-expected.txt


FAIL Property box-decoration-break has initial value slice assert_true: box-decoration-break doesn't seem to be supported in the computed style expected true got false
FAIL Property box-decoration-break does not inherit assert_true: expected true got false
PASS Property break-after has initial value auto
PASS Property break-after does not inherit
PASS Property break-before has initial value auto
PASS Property break-before does not inherit
PASS Property break-inside has initial value auto
PASS Property break-inside does not inherit
FAIL Property orphans has initial value 2 assert_equals: expected "2" but got "auto"
PASS Property orphans inherits
FAIL Property widows has initial value 2 assert_equals: expected "2" but got "auto"
PASS Property widows inherits
